Refactor compositor into modular sources
